refactor(client/echo): move getSpace to space-proxy.ts
getSpace lives next to SpaceProxy (it returns one) rather than in the
generic util.ts. util.ts now only contains the EDGE sync helpers.
