+++ packages/sdk/client/src/echo/util.ts
@@ -2,41 +2,8 @@
 // Copyright 2023 DXOS.org
 //
 
-import { type Space } from '@dxos/client-protocol';
-import { Obj } from '@dxos/echo';
 import { type SpaceSyncState } from '@dxos/echo-db';
 import { type SpaceId } from '@dxos/keys';
-
-import { SpaceProxy } from './space-proxy';
-
-/**
- * Returns the {@link Space} that owns the given object, or `undefined`.
- *
- * Use {@link Obj.getDatabase} when you only need DB/`spaceId` access; this
- * helper is retained only for callers that need {@link Space} proxy members
- * (`properties`, `queues`, `members`, `key`, `state`, `listen`, identity).
- */
-// TODO(burdon): Hypergraph.getSpace().
-export const getSpace = (object?: any): Space | undefined => {
-  if (!object) {
-    return undefined;
-  }
-
-  const db = Obj.getDatabase(object);
-  const id = db?.spaceId;
-  if (id && '_getOwningObject' in db.graph) {
-    const owner = (db.graph as { _getOwningObject: (id: SpaceId) => unknown })._getOwningObject(id);
-    if (owner instanceof SpaceProxy) {
-      return owner;
-    }
-  }
-
-  return undefined;
-};
-
-//
-// EDGE Sync State
-//
 
 export type PeerSyncState = Omit<SpaceSyncState.PeerState, 'peerId'>;
